restructured project to split api and db functions into their own files
